+++ front-end/e-maintenance-system/src/page/home/components/DrawerHome.js
@@ -1,0 +1,13 @@
+import { Divider, List } from "@material-ui/core";
+import React from "react";
+
+export default function DrawerHome() {
+    return (
+        <div>
+            <Divider />
+            <List>Home</List>
+            <Divider />
+            <List>About us</List>
+        </div>
+    );
+}
