+++ front-end/e-maintenance-system/src/routers/contants.js
@@ -1,0 +1,8 @@
+export const LIST_ROUTE = {
+    HOME_PAGE: "/",
+    LOGIN_PAGE: "/login",
+    ADMIN_PAGE: "/:adminID/admin",
+    BLOG: "/blog",
+    CONTACT: "/contact",
+    ABOUT_US: "/about",
+};
